requirements.txt
psycopg2-binary==2.9.10
streamlit==1.39.0
python-dotenv==1.0.1
docling==2.3.1
llama-index==0.11.21
llama-index-vector-stores-postgres==0.2.6
llama-index-embeddings-huggingface==0.3.1
llama-index-readers-docling==0.2.1
llama-index-node-parser-docling==0.2.0